[@scope] Semicolons should not abort selector parsing for @scope
When parsing rules under CSSNestingType::kScope, we always have some
non-nullptr parent_rule_for_nesting. This would activate special error
handling of semicolons (see AbortsNestedSelectorParsing) within
@scope, which (at least for now) doesn't support declarations directly
within its body (unless it's a nested group rule [1]):

  @scope (div) {
    z-index: 1;
  }

For the above (invalid) case, we'd parse the body using CSSParserImpl::
ConsumeRuleList, which would try to parse 'z-index:...' as a selector
via (ConsumeQualifiedRule). We would then abort at the semicolon,
and then eventually return to ConsumeRuleList, which would re-enter
ConsumeQualifiedRule and immediately abort at the semicolon (and so on).

This CL fixes this infinite loop by adding a separate flag for the
semicolon handling. This makes it possible to parse a list of rules
under CSSNestingType::kScope without also getting that handling.

Also added a DCHECK to make it easier to catch infinite loops
in the future.

Bug: 1456435

[1] https://drafts.csswg.org/css-nesting-1/#nested-group-rules

Change-Id: I9660e98343bdd47522fe87296b1da7112800945a