[@scope] Support @scope nested within style rules
This CL adds support for nesting @scope inside regular style rules, e.g.

  .a {
    @scope (& .b) {
      & { ... }
    }
  }

Note that @scope is a little different from other nested group rules
(like @media), because @scope itself acts as the parent *style* rule
for the purposes of evaluating the parent pseudo-class (&) inside the
body [1].

This means the '&' in the @scope's prelude refers to the '.a' selector,
but the '&' in @scope's body refers to the selector in the prelude,
which effectively is '.a .b'.

[1] https://drafts.csswg.org/css-nesting/#nesting-at-scope

Bug: 1456435
Change-Id: I6f53566d9e9bde50a888ab3b86dfd6699f004ebd