Safelist ftp/ftps/sftp for registerProtocolHandler
Adds "ftp", "ftps" and "sftp" to the safe list of protocols for
registerProcolHandler().

These schemes are already part of the spec
https://html.spec.whatwg.org/multipage/system-state.html#safelisted-scheme

It mitigates the risk of credential leakege for standard schemes by
removing the credentials from the url. See
https://html.spec.whatwg.org/multipage/system-state.html#security-and-privacy

This CL also updates the corresponding Web Platform Test, adding these
schemes to the safelist to comply with the spec.

Bug: 1199027
Change-Id: Ia68e7f7f8877aa00429519936195a1f690d9e603