[@scope] Improve sibling invalidation
ExtractInvalidationSetFeaturesFromCompound is supposed to return
the last simple selector of the compound, but did not account for
the kScopeActivation relation, which acts identically to kSubSelector
in this situation. This would cause unexpected behavior for selectors
with :scope (and therefore a kScopeActivation) in the subject, since
we would not be on a "true" combinator boundary after the call to
ExtractInvalidationSetFeaturesFromCompound.

Bug: 1451091
Change-Id: I550ccf87c54c0202f2aa46c856f1a53f218b783f